Crate rclrust

Re-exports

pub use clock::Clock;

pub use context::Context;

pub use init_options::InitOptions;

pub use log::Logger;

pub use node_options::NodeOptions;

pub use parameter::Parameter;

pub use time::Time;

pub use utility::*;

pub use qos::QoSProfile;

Modules

clock

context

error

init_options

log

node

node_options

parameter

qos

QoS (Quality of Service)

time

timer

utility

Macros

rclrust_debug

rclrust_error

rclrust_fatal

rclrust_info

rclrust_warn

Structs

Client

Service client

ParameterValue

Publisher

Publisher

Service

Service server

Subscription

Subscription

Enums

ClockType

Time source type, used to indicate the source of a time measurement.

DurabilityPolicy

QoS durability enumerations describing how samples persist

HistoryPolicy

QoS history enumerations describing how samples endure

LivelinessPolicy

QoS liveliness enumerations that describe a publisher’s reporting policy for its alive status. For a subscriber, these are its requirements for its topic’s publishers.

ParameterType