Struct rclrust::qos::QoSProfile

pub struct QoSProfile {
    pub history: HistoryPolicy,
    pub depth: usize,
    pub reliability: ReliabilityPolicy,
    pub durability: DurabilityPolicy,
    pub deadline: Duration,
    pub lifespan: Duration,
    pub liveliness: LivelinessPolicy,
    pub liveliness_lease_duration: Duration,
    pub avoid_ros_namespace_conventions: bool,
}

QoS profile

Fields
history: HistoryPolicy

History QoS policy setting.

depth: usize

Size of the message queue.

reliability: ReliabilityPolicy

Reliabiilty QoS policy setting.

durability: DurabilityPolicy

Durability QoS policy setting.

deadline: Duration

The period at which messages are expected to be sent/received.

lifespan: Duration

The age at which messages are considered expired and no longer valid.

liveliness: LivelinessPolicy

Liveliness QoS policy setting.

liveliness_lease_duration: Duration

The time within which the RMW node or publisher must show that it is alive.

avoid_ros_namespace_conventions: bool

If true, any ROS specific namespacing conventions will be circumvented.

In the case of DDS and topics, for example, this means the typical ROS specific prefix of rt would not be applied as described here:

http://design.ros2.org/articles/topic_and_service_names.html#ros-specific-namespace-prefix

This might be useful when trying to directly connect a native DDS topic with a ROS 2 topic.

Implementations

impl QoSProfile

pub const fn sensor_data() -> Self

Sensor Data QoS class

History: Keep last,

Depth: 5,

Reliability: Best effort,

Durability: Volatile,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

avoid ros namespace conventions: false

pub const fn parameters() -> Self

Parameters QoS class

History: Keep last,

Depth: 1000,

Reliability: Reliable,

Durability: Volatile,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn default() -> Self

Default QoS class

History: Keep last,

Depth: 10,

Reliability: Reliable,

Durability: Volatile,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn services_default() -> Self

Services QoS class

History: Keep last,

Depth: 10,

Reliability: Reliable,

Durability: Volatile,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn parameter_events() -> Self

Parameter events QoS class

History: Keep last,

Depth: 1000,

Reliability: Reliable,

Durability: Volatile,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn system_default() -> Self

System defaults QoS class

History: System default,

Depth: System default,

Reliability: System default,

Durability: System default,

Deadline: Default,

Lifespan: Default,

Liveliness: System default,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn unknown() -> Self

Unknow QoS class

History: Unknown,

Depth: System default,

Reliability: Unknown,

Durability: Unknown,

Deadline: Default,

Lifespan: Default,

Liveliness: Unknown,

Liveliness lease duration: Default,

Avoid ros namespace conventions: false

pub const fn history(self, history: HistoryPolicy) -> Self

Set the history policy.

Examples

let qos = QoSProfile::default().history(HistoryPolicy::KeepAll);

pub const fn keep_last(self, depth: usize) -> Self

Set the history to keep last.

Examples

let qos = QoSProfile::default().keep_last(10);

pub const fn keep_all(self) -> Self

Set the history to keep all.

Examples

let qos = QoSProfile::default().keep_all();

pub const fn reliability(self, reliability: ReliabilityPolicy) -> Self

Set the reliability setting.

Examples

let qos = QoSProfile::default().reliability(ReliabilityPolicy::Reliable);

pub const fn reliable(self) -> Self

Set the reliability setting to reliable.

Examples

let qos = QoSProfile::default().reliable();

pub const fn best_effort(self) -> Self

Set the reliability setting to best effort.

Examples

let qos = QoSProfile::default().best_effort();

pub const fn durability(self, durability: DurabilityPolicy) -> Self

Set the durability setting.

Examples

let qos = QoSProfile::default().durability(DurabilityPolicy::Volatile);

pub const fn volatile(self) -> Self

Set the durability setting to volatile.

Examples

let qos = QoSProfile::default().volatile();

pub const fn transient_local(self) -> Self

Set the durability setting to transient local.

Examples

let qos = QoSProfile::default().transient_local();

pub const fn deadline(self, deadline: Duration) -> Self

Set the deadline setting.

Examples

let qos = QoSProfile::default().deadline(Duration::from_secs(5));

pub const fn lifespan(self, lifespan: Duration) -> Self

Set the lifespan setting.

Examples

let qos = QoSProfile::default().lifespan(Duration::from_secs(5));

pub const fn liveliness(self, liveliness: LivelinessPolicy) -> Self

Set the liveliness setting.

Examples

let qos = QoSProfile::default().liveliness(LivelinessPolicy::Automatic);

pub const fn liveliness_lease_duration(
self,
liveliness_lease_duration: Duration
) -> Self

Set the liveliness_lease_duration setting.

Examples

let qos = QoSProfile::default().liveliness_lease_duration(Duration::from_secs(5));

pub const fn avoid_ros_namespace_conventions(
self,
avoid_ros_namespace_conventions: bool
) -> Self

Set the avoid_ros_namespace_conventions setting.

Examples

let qos = QoSProfile::default().avoid_ros_namespace_conventions(true);